Refactored, Riemersma dithering.
