+++ backend/src/nodes/impl/dithering/__init__.py
@@ -1,0 +1,8 @@
+
+
+# TODO handle palettes
+#  extract from image
+#  optimal palette selection
+#  ensure palettes with a different colorspace from the image work (just Greysclae/RGB)
+#  support different color distance functions
+# TODO hilbert curve diffusion
