+++ backend/src/nodes/nodes/image_adjustment/dither.py
@@ -6,12 +6,13 @@ from enum import Enum
 from sanic.log import logger
 
 from . import category as ImageAdjustmentCategory
-from ...impl.dithering import (uniform_quantize, ordered_dither, error_diffusion_dither,
-                               ThresholdMap, THRESHOLD_MAP_LABELS,
-                               ErrorDiffusionMap, ERROR_PROPAGATION_MAP_LABELS)
+from ...impl.dithering.common import uniform_quantize
+from ...impl.dithering.ordered import ThresholdMap, THRESHOLD_MAP_LABELS, ordered_dither
+from ...impl.dithering.diffusion import error_diffusion_dither, ErrorDiffusionMap, ERROR_PROPAGATION_MAP_LABELS
+from ...impl.dithering.riemersma import riemersma_dither
 from ...node_base import NodeBase, group
 from ...node_factory import NodeFactory
-from ...properties.inputs import ImageInput, NumberInput, EnumInput
+from ...properties.inputs import ImageInput, NumberInput, EnumInput, SliderInput
 from ...properties.outputs import ImageOutput
 
 
@@ -19,17 +20,19 @@ class UniformDitherAlgorithm(Enum):
     NONE = "None"
     ORDERED = "Ordered"
     DIFFUSION = "Diffusion"
+    RIEMERSMA = "Riemersma"
 
 
 UNIFORM_DITHER_ALGORITHM_LABELS = {
     UniformDitherAlgorithm.NONE: "No dithering",
     UniformDitherAlgorithm.ORDERED: "Ordered Dithering",
     UniformDitherAlgorithm.DIFFUSION: "Error Diffusion",
+    UniformDitherAlgorithm.RIEMERSMA: "Riemersma Diffusion",
 }
 
 
 @NodeFactory.register("chainner:image:dither")
-class OitherNode(NodeBase):
+class DitherNode(NodeBase):
     def __init__(self):
         super().__init__()
         self.description = "Apply one of a variety of dithering algorithms with a uniform (evenly-spaced) palette."
@@ -43,28 +46,25 @@ class OitherNode(NodeBase):
                 "conditional-enum",
                 {
                     "enum": 2,
-                    "conditions": [UniformDitherAlgorithm.ORDERED.value],
+                    "conditions": [UniformDitherAlgorithm.ORDERED.value,
+                                   UniformDitherAlgorithm.DIFFUSION.value,
+                                   UniformDitherAlgorithm.RIEMERSMA.value],
                 },
             )(
                 EnumInput(
                     ThresholdMap,
                     option_labels=THRESHOLD_MAP_LABELS,
                     default_value=ThresholdMap.BAYER_16).with_id(3),
-            ),
-
-            group(
-                "conditional-enum",
-                {
-                    "enum": 2,
-                    "conditions": [UniformDitherAlgorithm.DIFFUSION.value],
-                },
-            )(
                 EnumInput(
                     ErrorDiffusionMap,
                     option_labels=ERROR_PROPAGATION_MAP_LABELS,
                     default_value=ErrorDiffusionMap.FLOYD_STEINBERG).with_id(4),
+                NumberInput(
+                    "History Length",
+                    minimum=2,
+                    default=16,
+                ).with_id(5),
             )
-
         ]
         self.outputs = [ImageOutput()]
         self.category = ImageAdjustmentCategory
@@ -75,7 +75,8 @@ class OitherNode(NodeBase):
     def run(self, img: np.ndarray, num_colors: int,
             dither_algorithm: UniformDitherAlgorithm,
             threshold_map: ThresholdMap,
-            error_diffusion_map: ErrorDiffusionMap
+            error_diffusion_map: ErrorDiffusionMap,
+            history_length: int,
             ) -> np.ndarray:
         if dither_algorithm == UniformDitherAlgorithm.NONE:
             return uniform_quantize(img, num_colors=num_colors)
@@ -83,6 +84,8 @@ class OitherNode(NodeBase):
             return ordered_dither(img, num_colors=num_colors, threshold_map=threshold_map)
         elif dither_algorithm == UniformDitherAlgorithm.DIFFUSION:
             return error_diffusion_dither(img, num_colors=num_colors, error_diffusion_map=error_diffusion_map)
+        elif dither_algorithm == UniformDitherAlgorithm.RIEMERSMA:
+            return riemersma_dither(img, num_colors=num_colors, history_length=history_length, decay_ratio=1/history_length)
 
 
 class PaletteDitherAlgorithm(Enum):
@@ -96,7 +99,7 @@ PALETTE_DITHER_ALGORITHM_LABELS = {
 }
 
 # @NodeFactory.register("chainner:image:palette_dither")
-# class PaletteOitherNode(NodeBase):
+# class PaletteDitherNode(NodeBase):
 #     def __init__(self):
 #         super().__init__()
 #         self.description = "Apply one of a variety of dithering algorithms with a provided palette.  A palette is just an image with one row where each pixel is a color in the palette."
